pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ppt-maker"
version = "0.1.0"
description = "범용 보고서 + 프레젠테이션 자동 생성 파이프라인"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "ppt-maker contributors" }]

dependencies = [
    "python-pptx>=1.0.0",
    "pptx-ea-font>=0.0.1",
    "Jinja2>=3.1",
    "typer>=0.9",
    "rich>=13.0",
    "tomli>=2.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
ppt-maker = "ppt_maker.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ppt_maker"]

[tool.ruff]
target-version = "py310"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008"]  # typer.Option/Argument in function defaults is standard Typer pattern

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
